1. Make your element visible within the New Content Element Wizard (NCEW)
We add a new element to the wizard with TSConfig

mod.wizards.newContentElement.wizardItems
includes all groups and items of the wizard.
https://docs.typo3.org/typo3cms/TSconfigReference/PageTsconfig/Mod/Index.html?highlight=wizarditems#new-content-element-wizard-mod-wizards-newcontentelement

line 2) We create a new group with the identifier "homepage"
line 3) We set the header of this group
line 4) We add all, instead of individually adding elements of this group
line 5) "elements" holds all our items of this group
line 6) We define a new identifier, which can be named as you like
line 7) The icon identifier of an already existing icon
line 8 and 9 are self-explanatory
line 10ff) all pre filled fields of your new record must be declared within "tt_content_defValues" array.
line 11) We only define the CType value "theme_tuga" (which will be later on our CType identifier)

The new element must be appear within the NCEW (after a clearing your cache)

¯\_(ツ)_/¯